+++ src/main/kotlin/team/gsm/flooding/domain/homebase/dto/response/AttendanceResponse.kt
@@ -1,4 +1,4 @@
-package team.gsm.flooding.domain.attendance.dto.response
+package team.gsm.flooding.domain.homebase.dto.response
 
 import team.gsm.flooding.domain.attendance.entity.Attendance
 import team.gsm.flooding.global.util.StudentUtil
@@ -8,14 +8,13 @@ class AttendanceResponse(
 	val schoolNumber: String,
 ) {
 	companion object {
-		fun toDto(attendance: Attendance): AttendanceResponse {
-			return AttendanceResponse(
+		fun toDto(attendance: Attendance): AttendanceResponse =
+			AttendanceResponse(
 				name = attendance.student.name,
 				schoolNumber =
 					attendance.student.studentInfo.let {
 						StudentUtil.calcStudentNumber(it.year, it.classroom, it.number)
 					},
 			)
-		}
 	}
 }
