+++ src/main/kotlin/team/gsm/flooding/domain/homebase/usecase/FindMyReservedHomebaseUsecase.kt
@@ -1,9 +1,9 @@
-package team.gsm.flooding.domain.attendance.usecase
+package team.gsm.flooding.domain.homebase.usecase
 
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import team.gsm.flooding.domain.attendance.dto.response.FindMyReservedHomebaseResponse
-import team.gsm.flooding.domain.attendance.repository.HomebaseGroupRepository
+import team.gsm.flooding.domain.homebase.dto.response.FindMyReservedHomebaseResponse
+import team.gsm.flooding.domain.homebase.repository.HomebaseGroupRepository
 import team.gsm.flooding.global.util.UserUtil
 import java.time.LocalDate
